Move file_contains function to common
The function is going to be needed by shebang check as well to work
with log file.
