+++ taskotron_python_versions/python_usage.py
@@ -1,6 +1,4 @@
-import mmap
-
-from .common import log, write_to_artifact
+from .common import log, write_to_artifact, file_contains
 
 
 MESSAGE = """You've used /usr/bin/python during build on the following arches:
@@ -14,30 +12,6 @@ Use /usr/bin/python3 or /usr/bin/python2 explicitly.
 INFO_URL = ('https://fedoraproject.org/wiki/Changes/'
             'Avoid_usr_bin_python_in_RPM_Build')
 WARNING = 'DEPRECATION WARNING: python2 invoked with /usr/bin/python'
-
-
-def file_contains(path, needle):
-    """Check if the file residing on the given path contains the given needle
-    """
-    # Since we have no idea if build.log is valid utf8, let's convert our ASCII
-    # needle to bytes and use bytes everywhere.
-    # This also allow us to use mmap on Python 3.
-    # Be explicit here, to make it fail early if our needle is not ASCII.
-    needle = needle.encode('ascii')
-
-    with open(path, 'rb') as f:
-        # build.logs tend to be laaaarge, so using a single read() is bad idea;
-        # let's optimize prematurely because practicality beats purity
-
-        # Memory-mapped file object behaving like bytearray
-        mmf = mmap.mmap(f.fileno(),
-                        length=0,  # = determine automatically
-                        access=mmap.ACCESS_READ)
-        try:
-            return mmf.find(needle) != -1
-        finally:
-            # mmap context manager is Python 3 only
-            mmf.close()
 
 
 def task_python_usage(logs, koji_build, artifact):
